backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.2
playwright==1.49.1
openai>=1.30.0
Pillow==11.1.0
modal==0.73.45
python-dotenv==1.0.1
httpx==0.28.1
aiohttp>=3.9.0
dnspython>=2.4.0
certifi>=2023.0.0
# Database & caching
asyncpg>=0.29.0
pydantic>=2.0.0
upstash-redis>=1.0.0
redis>=5.0.0
# embeddings use openai (already listed above)
playwright-stealth>=1.0.6
